requirements.txt
celery==5.2.6
django-celery-beat==2.2.1
django==3.2
djangorestframework==3.13.1
djangorestframework-xml==2.0.0
djangorestframework-simplejwt==5.1.0
fake-useragent==0.1.11
psycopg2-binary==2.9.3
requests==2.27.1
redis==4.3.1
